Heavy refactoring to display logic. Light refactoring elsewhere. Organized files and commented some code.
