+++ Droar.podspec
@@ -8,7 +8,7 @@
 
 Pod::Spec.new do |s|
   s.name             = 'Droar'
-  s.version          = '1.5.4'
+  s.version          = '2.0.0'
   s.summary          = 'A runtime debugging tool for iOS.'
 
   s.description      = <<-DESC
