+++ Droar/Classes/Extensions/UIColorExtensions.swift
@@ -8,15 +8,17 @@
 import Foundation
 
 internal extension UIColor {
+    
     static var droarBlue: UIColor {
-        get { return UIColor(red: 45.0 / 255, green: 88.0 / 255, blue: 124.0/255, alpha: 1.0) }
+        return UIColor(red: 45.0 / 255, green: 88.0 / 255, blue: 124.0/255, alpha: 1.0)
     }
     
     static var droarGray: UIColor {
-        get { return UIColor(red: 205.0 / 255, green: 207.0 / 255, blue: 209.0/255, alpha: 1.0) }
+        return UIColor(red: 205.0 / 255, green: 207.0 / 255, blue: 209.0/255, alpha: 1.0)
     }
     
     static var disabledGray: UIColor {
-        get { return UIColor(white: 0.95, alpha: 1.0)}
+        return UIColor(white: 0.95, alpha: 1.0)
     }
+    
 }
