+++ Droar/Classes/Extensions/UIFontExtensions.swift
@@ -10,6 +10,8 @@ import Foundation
 extension String: Error {}
 
 internal extension UIFont {
+    
+    //Registration
     static func register(url: URL) throws {
         guard let fontDataProvider = CGDataProvider(url: url as CFURL) else {
             throw "Could not create font data provider for \(url)."
@@ -26,4 +28,30 @@ internal extension UIFont {
             throw error!.takeUnretainedValue()
         }
     }
+    
+    static func font(named: String, size: CGFloat, resource: String) -> UIFont? {
+        //Get if the font exists
+        if let font = UIFont(name: named, size: size) { return font }
+        
+        //Attempt register if it does not
+        if let fontUrl = Bundle.podBundle.url(forResource: resource, withExtension: ".ttf") {
+            do {
+                try UIFont.register(url: fontUrl)
+            } catch let error {
+                print(error.localizedDescription)
+            }
+        }
+        
+        //Return new font if it exists
+        return UIFont(name: named, size: size)
+    }
+    
+    //Fonts
+    static func russo(_ size: CGFloat) -> UIFont {
+        let fontName = "Russo One"
+        let resourceName = "RussoOne-Regular"
+        
+        return UIFont.font(named: fontName, size: size, resource: resourceName) ?? UIFont.systemFont(ofSize: size)
+    }
+    
 }
