+++ Example/Droar/Base.lproj/Main.storyboard
@@ -1,12 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="13771" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="vXZ-lx-hvc">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14868" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="vXZ-lx-hvc">
+    <device id="retina4_7" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13772"/>
-        <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14824"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <customFonts key="customFonts">
@@ -28,7 +25,7 @@
                         <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
                         <subviews>
                             <imageView userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="DroarIcon_Text" translatesAutoresizingMaskIntoConstraints="NO" id="vIY-zf-O7Y">
-                                <rect key="frame" x="36" y="64" width="303" height="539"/>
+                                <rect key="frame" x="36" y="64.5" width="303" height="538"/>
                             </imageView>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" verticalCompressionResistancePriority="751" text="Swipe from far right to start!" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="kJo-5W-p1a">
                                 <rect key="frame" x="39.5" y="622.5" width="296.5" height="24.5"/>
@@ -36,13 +33,22 @@
                                 <color key="textColor" red="0.1764705882352941" green="0.34509803921568627" blue="0.48627450980392156" alpha="1" colorSpace="calibratedRGB"/>
                                 <nil key="highlightedColor"/>
                             </label>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="9JO-eT-kog">
+                                <rect key="frame" x="143.5" y="80" width="88" height="30"/>
+                                <state key="normal" title="Present New"/>
+                                <connections>
+                                    <action selector="presentTest:" destination="vXZ-lx-hvc" eventType="touchUpInside" id="arq-Ve-ncU"/>
+                                </connections>
+                            </button>
                         </subviews>
                         <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                         <constraints>
                             <constraint firstItem="vIY-zf-O7Y" firstAttribute="centerY" secondItem="kh9-bI-dsS" secondAttribute="centerY" id="0hF-3Y-nxc"/>
                             <constraint firstItem="kJo-5W-p1a" firstAttribute="top" secondItem="vIY-zf-O7Y" secondAttribute="bottom" constant="20" id="77f-7F-B61"/>
+                            <constraint firstItem="9JO-eT-kog" firstAttribute="top" secondItem="jyV-Pf-zRb" secondAttribute="bottom" constant="80" id="9jD-fb-xfD"/>
                             <constraint firstItem="vIY-zf-O7Y" firstAttribute="leading" secondItem="kh9-bI-dsS" secondAttribute="leadingMargin" constant="20" id="Qgq-UM-Vcm"/>
                             <constraint firstAttribute="trailingMargin" secondItem="vIY-zf-O7Y" secondAttribute="trailing" constant="20" id="XVY-yv-eZ7"/>
+                            <constraint firstItem="9JO-eT-kog" firstAttribute="centerX" secondItem="kh9-bI-dsS" secondAttribute="centerX" id="cuN-Y7-LqX"/>
                             <constraint firstItem="kJo-5W-p1a" firstAttribute="centerX" secondItem="kh9-bI-dsS" secondAttribute="centerX" id="iHu-kN-res"/>
                             <constraint firstItem="2fi-mo-0CV" firstAttribute="top" secondItem="kJo-5W-p1a" secondAttribute="bottom" constant="20" id="oxW-8a-JBg"/>
                         </constraints>
@@ -50,6 +56,7 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="x5A-6p-PRh" sceneMemberID="firstResponder"/>
             </objects>
+            <point key="canvasLocation" x="138" y="134"/>
         </scene>
     </scenes>
     <resources>
